feat: Added hot swap methods to Files class
